Moved status api to server/grpc/admin/status. Needed to move some functions out for streaming logs and scaning output
