+++ router/admin/ocelot_server.go
@@ -3,11 +3,12 @@ package admin
 import (
 	"github.com/level11consulting/ocelot/repo"
 	"github.com/level11consulting/ocelot/secret"
+	"github.com/level11consulting/ocelot/server/grpc/admin/status"
 )
 
 type OcelotServer interface {
 	secret.SecretInterface
 	repo.RepoInterface
-	StatusInterface
+	status.StatusInterface
 	BuildInterface
 }
